Conflicts:
	fact-team.github.io

+++ src/database/billingdatabase.cpp
@@ -38,10 +38,10 @@ Models::Billing* BillingDatabase::getBilling(const int pId) {
         billing->setTitle(value(q, "title").toString());
         billing->setDescription(value(q,"description").toString());
         billing->setNumber(value(q,"number").toInt());
-        billing->setDate(QDate::fromString(value(q,"date").toString(),"yyyy-MM-dd"));
+        billing->setDate(value(q,"date").toDate());
         billing->setIsBilling(value(q,"isBilling").toBool());
+        billing->setIsPaid(value(q,"isPaid").toBool());
         billing->setToRemoved(false);
-
     } else {
         billing = NULL;
     }
@@ -56,7 +56,7 @@ WdgModels::BillingsTableModel* BillingDatabase::getBillingsTable(
     QSqlQuery q;
     q.prepare(
              "SELECT DISTINCT b.idBilling, title, description, number, "
-             "isBilling, date "
+             "isBilling, date, isPaid "
              "FROM Billing b, BillingProject bp "
              "WHERE idProject = :idproject "
              "AND b.idBilling = bp.idBilling ORDER BY date DESC");
@@ -83,9 +83,9 @@ int BillingDatabase::addBilling(const Models::Billing& pBilling) {
     QSqlQuery q;
 
     q.prepare(  "INSERT INTO Billing "
-                "(title, description, number, isBilling, date)"
+                "(title, description, number, isBilling, date, isPaid)"
                 " VALUES "
-                "(:title, :description, :number, :isBilling, :date)"
+                "(:title, :description, :number, :isBilling, :date, :isPaid)"
                 );
 
     q.bindValue(":title", pBilling.getTitle());
@@ -93,7 +93,7 @@ int BillingDatabase::addBilling(const Models::Billing& pBilling) {
     q.bindValue(":number", pBilling.getNumber());
     q.bindValue(":isBilling", pBilling.isBilling());
     q.bindValue(":date", pBilling.getDate());
-
+    q.bindValue(":isPaid", pBilling.isPaid());
     if(!q.exec()) {
         throw new DbException(
                     "Impossible d'ajouter le Billing",
@@ -104,7 +104,9 @@ int BillingDatabase::addBilling(const Models::Billing& pBilling) {
     return q.lastInsertId().toInt();
 }
 
-void BillingDatabase::addBillingProject(const int idProject, const int idBilling, const int idContributory) {
+void BillingDatabase::addBillingProject(
+        const int idProject, const int idBilling, const int idContributory)
+{
     QSqlQuery q;
     q.prepare(
                 "INSERT INTO BillingProject "
@@ -124,12 +126,14 @@ void BillingDatabase::addBillingProject(const int idProject, const int idBilling
 
 }
 
-void BillingDatabase::removeBillingProject(const int idProject, const int idBilling, const int idContributory)
+void BillingDatabase::removeBillingProject(
+        const int idProject, const int idBilling, const int idContributory)
 {
     QSqlQuery q;
     QString project;
 
-    (idProject == 0) ? project="" : project="idProject= "+ QString::number(idProject) +" AND ";
+    (idProject == 0) ? project="" :
+            project="idProject= "+ QString::number(idProject) +" AND ";
 
     q.prepare(
                 "DELETE FROM BillingProject "
@@ -156,7 +160,8 @@ void BillingDatabase::updateBilling(const Models::Billing& pBilling)
              "title=:title, "
              "description=:description, "
              "number=:number, "
-             "date=:date "
+             "date=:date, "
+             "isPaid=:isPaid "
              "WHERE idBilling=:idBilling"
              );
 
@@ -166,6 +171,7 @@ void BillingDatabase::updateBilling(const Models::Billing& pBilling)
    //q.bindValue(":isBilling", pBilling.isBilling());
    q.bindValue(":date", pBilling.getDate());
    q.bindValue(":idBilling",pBilling.getId());
+   q.bindValue(":isPaid", pBilling.isPaid());
 
    if(!q.exec()) {
        throw new DbException(
@@ -175,38 +181,78 @@ void BillingDatabase::updateBilling(const Models::Billing& pBilling)
                    1.4);
    }
 }
+bool BillingDatabase::isBillingPaid(const int pId) {
+    QSqlQuery q;
+
+    q.prepare("SELECT count(*) AS nb_b "
+              "FROM Billing "
+              "WHERE idBilling = :pId "
+              "AND isBilling = 1 "
+              "AND isPaid = 1 ");
+
+    q.bindValue(":pId",pId);
+
+    if(!q.exec()) {
+        throw new DbException(
+                    "Impossible de supprimer le Billing ",
+                    "BddContributory::removeBilling",
+                    lastError(q),
+                    1.5);
+    }
+
+    q.next();
+
+     return value(q, "nb_b").toInt() == 1;
+
+}
 
 void BillingDatabase::removeBilling(const int pId)
 {
     QSqlQuery q;
-    q.prepare("DELETE FROM BillingProject "
-              "WHERE idBilling=:pId");
 
-    q.bindValue(":pId",pId);
+    if (!isBillingPaid(pId)) {
+        q.prepare("DELETE FROM BillingProject "
+                  "WHERE idBilling=:pId");
 
-    if(!q.exec()) {
-        throw new DbException(
-                    "Impossible de supprimer le Billing ",
-                    "BddContributory::removeBilling",
-                    lastError(q),
-                    1.5);
+        q.bindValue(":pId",pId);
+
+        if(!q.exec()) {
+            throw new DbException(
+                        "Impossible de supprimer le Billing ",
+                        "BddContributory::removeBilling",
+                        lastError(q),
+                        1.51);
+        }
+        q.prepare("DELETE FROM BillingRate "
+                  "WHERE idBilling=:pId");
+
+        q.bindValue(":pId",pId);
+
+        if(!q.exec()) {
+            throw new DbException(
+                        "Impossible de supprimer le Billing ",
+                        "BddContributory::removeBilling",
+                        lastError(q),
+                        1.51);
+        }
+
+        q.clear();
+
+        q.prepare(
+                    "DELETE FROM Billing "
+                    "WHERE idBIlling=:pId");
+
+        q.bindValue(":pId",pId);
+
+        if(!q.exec()) {
+            throw new DbException(
+                        "Impossible de supprimer le Billing ",
+                        "BddContributory::removeBilling",
+                        lastError(q),
+                        1.52);
+        }
     }
 
-    q.clear();
-
-    q.prepare(
-                "DELETE FROM Billing "
-                "WHERE idBIlling=:pId");
-
-    q.bindValue(":pId",pId);
-
-    if(!q.exec()) {
-        throw new DbException(
-                    "Impossible de supprimer le Billing ",
-                    "BddContributory::removeBilling",
-                    lastError(q),
-                    1.5);
-    }
 }
 int BillingDatabase::getMaxBillingNumber()
 {
@@ -239,6 +285,55 @@ int BillingDatabase::getMaxQuoteNumber()
     return value(q, "max").toInt();
 }
 
+int BillingDatabase::getMaxBillingNumberOfCustomer(const int idCustomer)
+{
+    QSqlQuery q;
+    q.prepare("SELECT MAX(number) as max FROM Customer c, Project p, BillingProject bp, Billing b "
+              "WHERE "
+              "c.idCustomer = p.idCustomer "
+              "AND p.idProject = bp.idProject "
+              "AND bp.idBilling = b.idBilling "
+              "AND isBilling = 1 "
+              "AND c.idCustomer = :idCustomer");
+
+    q.bindValue(":idCustomer",idCustomer);
+
+    if(!q.exec()) {
+        throw new DbException(
+            "Impossible d'obtenir le numéro max du devis",
+            "BddContributory::getMaxBillingNumberOfCustomer",
+            lastError(q),
+            1.6);
+    }
+    q.first();
+    return value(q,"max").toInt();
+
+}
+
+int BillingDatabase::getMaxQuoteNumberOfCustomer(const int idCustomer)
+{
+    QSqlQuery q;
+    q.prepare("SELECT MAX(number) as max FROM Customer c, Project p, BillingProject bp, Billing b "
+              "WHERE "
+              "c.idCustomer = p.idCustomer "
+              "AND p.idProject = bp.idProject "
+              "AND bp.idBilling = b.idBilling "
+              "AND isBilling = 0 "
+              "AND c.idCustomer = :idCustomer ");
+
+    q.bindValue(":idCustomer",idCustomer);
+
+    if(!q.exec()) {
+        throw new DbException(
+            "Impossible d'obtenir le numéro max du devis",
+            "BddContributory::getMaxBillingNumberOfCustomer",
+            lastError(q),
+            1.6);
+    }
+    q.first();
+    return value(q,"max").toInt();
+}
+
 QSharedPointer<Billing> BillingDatabase::getBilling(QSqlQuery &q)
 {
     QSharedPointer<Models::Billing> billing =
@@ -247,8 +342,9 @@ QSharedPointer<Billing> BillingDatabase::getBilling(QSqlQuery &q)
     billing->setTitle(value(q, "title").toString());
     billing->setDescription(value(q,"description").toString());
     billing->setNumber(value(q,"number").toInt());
-    billing->setDate(QDate::fromString(value(q,"date").toString(),"yyyy-MM-dd"));
+    billing->setDate(value(q,"date").toDate());
     billing->setIsBilling(value(q,"isBilling").toBool());
+    billing->setIsPaid(value(q,"isPaid").toBool());
     billing->setToRemoved(false);
 
     return billing;
@@ -274,6 +370,97 @@ QMap<Project*, Billing*> BillingDatabase::getAllBillingsOfProject()
                     new Billing(value(q, "idBilling").toInt()));
     }
     return map;
+}
+
+QList<Billing> BillingDatabase::getBillings(const int projectId)
+{
+    QList<Billing> bills;
+    QSqlQuery q;
+    q.prepare(
+             "SELECT DISTINCT b.idBilling, title, description, number, "
+             "isBilling, date, isPaid "
+             "FROM Billing b, BillingProject bp "
+             "WHERE idProject = :idproject "
+             "AND b.idBilling = bp.idBilling AND b.isBilling = 1 "
+             "ORDER BY date DESC");
+
+    q.bindValue(":idproject",projectId);
+
+    if(!q.exec()) {
+        throw new DbException(
+                    "Impossible de récupérer les Factures",
+                    "BddCustomer::getBillings",
+                    lastError(q),
+                    1.3);
+    }
+
+    while(q.next()) {
+        bills.append(*getBilling(q));
+    }
+    return bills;
+}
+
+QList<Billing> BillingDatabase::getAllBillingsOnly(const int idProject)
+{
+    QList<Billing> bills;
+    QSqlQuery q;
+    q.prepare(
+             "SELECT DISTINCT b.idBilling, title, description, number, "
+             "isBilling, date, isPaid "
+             "FROM Billing b, BillingProject bp "
+             "WHERE idProject=:idProject "
+             "AND b.idBilling = bp.idBilling "
+             "AND isPaid = 1 "
+             "AND isBilling = 1 "
+             );
+
+    q.bindValue(":idProject",idProject);
+
+
+    if(!q.exec()) {
+        throw new DbException(
+                    "Impossible de récupérer les Factures",
+                    "BddCustomer::getBillings",
+                    lastError(q),
+                    1.3);
+    }
+
+    while(q.next()) {
+        bills.append(*getBilling(q));
+    }
+    return bills;
+}
+
+QList<Billing> BillingDatabase::getBillingsBetweenDates(QDate begin, QDate end)
+{
+    // select * from billing where date between '2015-01-01' and '2015-03-19' and isBilling=1 ;
+    QList<Billing> bills;
+    QSqlQuery q;
+    q.prepare(
+             "SELECT DISTINCT b.idBilling, title, description, number, "
+             "isBilling, date, isPaid "
+             "FROM Billing b "
+             "WHERE "
+              " date BETWEEN :begin and :end "
+              "AND isPaid = 1 "
+              "AND isBilling = 1 "
+             );
+    q.bindValue(":begin", begin);
+    q.bindValue(":end", end);
+    if(!q.exec()) {
+        throw new DbException(
+                    "Impossible de récupérer les Factures",
+                    "BddCustomer::getBillings",
+                    lastError(q),
+                    1.3);
+    }
+
+    while(q.next()) {
+        bills << *getBilling(q);
+    }
+
+    return bills;
+
 }
 
 }
